requirements.txt
matplotlib==3.5.2
numpy==1.21.6
pandas==1.3.2
streamlit==1.11.1
scipy==1.9.2
pillow==9.3